requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose[cryptography]
streamlit
requests
bcrypt>=4.0.0
